Add invite-send-location test

+++ db/reset.ts
@@ -4,7 +4,10 @@ import selectAllNonMetaTables from './selectAllNonMetaTables.ts'
 export default async function reset() {
   const tables = await selectAllNonMetaTables(db)
   for (const table of tables) {
-    console.log(`Deleting all rows from ${table}`)
+    if (table === 'facilities') {
+      continue
+    }
+    //console.log(`Deleting all rows from ${table}`)
     // deno-lint-ignore no-explicit-any
     await db.deleteFrom(table as any).execute()
   }
